+++ CHANGELOG.rst
@@ -10,6 +10,8 @@ Versioning <https://semver.org/spec/v2.0.0.html>`_.
 Unreleased_
 -----------
 
+- `Allow "status" and "severity" on In record init <../../pull/111>`_
+
 4.1.0_ - 2022-08-05
 -------------------